requirements.txt
asgiref==3.7.2
certifi==2023.11.17
click==8.1.7
cloudinary==1.37.0
colorama==0.4.6
cssbeautifier==1.14.11
Django==4.2.8
django-htmx==1.17.2
django-localflavor==4.0
EditorConfig==0.12.3
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.11
json5==0.9.14
packaging==23.2
pathspec==0.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-stdnum==1.19
PyYAML==6.0.1
regex==2023.12.25
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
urllib3==1.26.16
